fix --no-verify
the --no-verify flag has been broken since Python 3.4
when it is used together with --allow-insecure-ciphers.

this is because setting cert_mode to NONE now requires
previously setting check_hostname to False [1].

again, urllib3 normally takes care of this [2], but because
we set our own ssl_context, we need to replicate this logic.

[1]: https://github.com/python/cpython/commit/1aa9a75fbff2333fd07574e3de8710c629483258#diff-89879be484d86da4e77c90d5408b2e10190ee27cc86337cd0f8efc3520d60621R2237-R2242
[2]: https://github.com/urllib3/urllib3/blob/b7c2910a9f49ba0d58c8d4381500c208f9a24765/src/urllib3/util/ssl_.py#L327-L337